ネットワーク図作成処理をhooksに切り出した
