+++ src/hooks/d3.d.ts
@@ -1,0 +1,1 @@
+declare module "d3";
